+++ src/main/java/webserver/requestline/HttpMethod.java
@@ -1,4 +1,4 @@
-package webserver.domain;
+package webserver.requestline;
 
 import exception.ResourceNotFound;
