+++ src/main/java/webserver/requestline/HttpVersion.java
@@ -1,4 +1,4 @@
-package webserver.domain;
+package webserver.requestline;
 
 import exception.ResourceNotFound;
