+++ src/main/java/webserver/requestline/Protocol.java
@@ -1,4 +1,4 @@
-package webserver.domain;
+package webserver.requestline;
 
 public enum Protocol {
