+++ src/test/java/webserver/requestline/HttpVersionTest.java
@@ -1,4 +1,4 @@
-package webserver.domain;
+package webserver.requestline;
 
 import exception.ResourceNotFound;
 import org.junit.jupiter.api.DisplayName;
